requirements.txt
asgiref==3.11.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.5
